Adding indentity server configuration to auction service

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -55,12 +56,13 @@ public class AuctionController : ControllerBase
         return _mapper.Map<AuctionDTO>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDTO>> CreateAuction(CreateAuctionDTO auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        // TODO: Add current user as seller
-        auction.Seller = "test";
+        
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
 
@@ -79,6 +81,7 @@ public class AuctionController : ControllerBase
             new { id = auction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDTO auctionDto)
     {
@@ -91,7 +94,10 @@ public class AuctionController : ControllerBase
             return NotFound();
         }
 
-        // TODO: check seller = username
+        if( auction.Seller != User.Identity.Name )
+        {
+            return Forbid();
+        }
 
         auction.Item.Make = auctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = auctionDto.Model ?? auction.Item.Model;
@@ -111,6 +117,7 @@ public class AuctionController : ControllerBase
         return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -120,7 +127,10 @@ public class AuctionController : ControllerBase
             return NotFound();
         }
 
-        // TODO: check seller = username
+        if( auction.Seller != User.Identity.Name )
+        {
+            return Forbid();
+        }
 
         _context.Auctions.Remove(auction);
